[fix] 방장 화면으로 전체 게임 화면 구현 + 게임방으로 돌아가기

+++ src/main/java/com/bettopia/game/socket/TurtleRunWebsocketHandler.java
@@ -1,13 +1,9 @@
 package com.bettopia.game.socket;
 
 import com.bettopia.game.model.auth.AuthService;
-import com.bettopia.game.model.auth.UserVO;
 import com.bettopia.game.model.gameroom.GameRoomDAO;
-import com.bettopia.game.model.gameroom.GameRoomResponseDTO;
 import com.bettopia.game.model.gameroom.GameRoomService;
-import com.bettopia.game.model.history.GameHistoryDTO;
 import com.bettopia.game.model.history.HistoryService;
-import com.bettopia.game.model.history.PointHistoryDTO;
 import com.bettopia.game.model.multi.turtle.PlayerDAO;
 import com.bettopia.game.model.multi.turtle.TurtlePlayerDTO;
 import com.bettopia.game.model.multi.turtle.TurtleRunResultDTO;
@@ -20,7 +16,6 @@ import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketSession;
 import org.springframework.web.socket.handler.TextWebSocketHandler;
-
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -49,8 +44,11 @@ public class TurtleRunWebsocketHandler extends TextWebSocketHandler {
     private AuthService authService;
     @Autowired
     private HistoryService historyService;
+    @Autowired
+	private GameRoomListWebSocket gameRoomListWebSocket;
+    
 
-    private final Map<String, Map<String, List<Double>>> latestPositions = new ConcurrentHashMap<>();
+    private final Map<String, List<Double>> latestPositions = new ConcurrentHashMap<>();
     private final Map<String, ScheduledFuture<?>> broadcastTasks = new ConcurrentHashMap<>();
     private final ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(4);
     
@@ -63,8 +61,6 @@ public class TurtleRunWebsocketHandler extends TextWebSocketHandler {
 	private void broadcastMessage(String type, String roomId, Map<String, Object> data) throws IOException {
 		// 웹소켓 메시지 전송
 		List<WebSocketSession> sessions = sessionService.getSessions(roomId);
-		List<TurtlePlayerDTO> players = playerDAO.getAll(roomId);
-
 		if (sessions == null || sessions.isEmpty()) {
 			// 더 이상 메시지 보낼 대상이 없음
 			return;
@@ -73,7 +69,6 @@ public class TurtleRunWebsocketHandler extends TextWebSocketHandler {
 		ObjectMapper mapper = new ObjectMapper();
 		Map<String, Object> messageMap = new HashMap<>();
 		messageMap.put("type", type);
-		messageMap.put("players", players);
 
 		if (data != null) {
 			messageMap.putAll(data);
@@ -114,9 +109,7 @@ public class TurtleRunWebsocketHandler extends TextWebSocketHandler {
                         positions.add(pos.asDouble());
                     }
                 }
-                latestPositions
-                .computeIfAbsent(roomId, k -> new ConcurrentHashMap<>())
-                .put(userId, positions);
+                latestPositions.put(roomId, positions);
 	            // 2. 전체 클라이언트에게 positions 전송
                 broadcastTasks.computeIfAbsent(roomId, k -> {
                     return scheduler.scheduleAtFixedRate(() -> {
@@ -125,7 +118,7 @@ public class TurtleRunWebsocketHandler extends TextWebSocketHandler {
                         } catch (IOException e) {
                             e.printStackTrace();
                         }
-                    }, 0, 100, TimeUnit.MILLISECONDS);
+                    }, 0, 20, TimeUnit.MILLISECONDS);
                 });
                 break;
 			case "race_finish":
@@ -151,46 +144,45 @@ public class TurtleRunWebsocketHandler extends TextWebSocketHandler {
 				data.put("target",  "/gameroom/detail/" + roomId);
 				broadcastMessage("end", roomId, data);
 		}
-
-		List<TurtlePlayerDTO> players = playerDAO.getAll(roomId);
-		Map<String, Object> data = new HashMap<>();
-		data.put("players", players);
-		
-		broadcastMessage("update", roomId, data);
 	}
 	
 	// 방에 위치 정보를 스케쥴러로 보내주는 함수
 	private void broadcastRaceUpdate(String roomId) throws IOException {
-	    Map<String, List<Double>> positionsByUser = latestPositions.get(roomId);
-	    if (positionsByUser == null) return;
+	    List<Double> positions = latestPositions.get(roomId);
+	    if (positions == null) return;
 
 	    Map<String, Object> data = new HashMap<>();
-	    data.put("positions", positionsByUser);
+	    data.put("positions", positions);
 	    broadcastMessage("race_update", roomId, data);
 	}
 	
 	@Override
 	public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
-		// 세션 제거, 퇴장 처리 등
-		String roomId = (String) session.getAttributes().get("roomId");
-		String userId = (String) session.getAttributes().get("userId");
+	    // 세션 제거, 퇴장 처리 등
+	    String roomId = (String) session.getAttributes().get("roomId");
+	    String userId = (String) session.getAttributes().get("userId");
 
-		sessionService.removeSession(roomId, session);
-		playerDAO.removePlayer(roomId, userId);
+	    sessionService.removeSession(roomId, session);
 
-		// 플레이어가 0명일 때 방 삭제
-		List<TurtlePlayerDTO> players = playerDAO.getAll(roomId);
-		if (players == null || players.isEmpty()) {
-			gameroomDAO.deleteRoom(roomId);
-			// 스케줄러도 종료
-			ScheduledFuture<?> task = broadcastTasks.remove(roomId);
+	    String gameStatus = gameRoomService.selectById(roomId).getStatus();
+	    if(!gameStatus.equals("WAITING")) {
+	       playerDAO.removePlayer(roomId, userId);
+	       gameRoomListWebSocket.broadcastMessage("exit");
+
+	       Map<String, Object> data = new HashMap<>();
+	       data.put("userId", userId);
+	       broadcastMessage("exit", roomId, data);
+	    }
+
+	    // 플레이어가 0명일 때 방 삭제
+	    List<TurtlePlayerDTO> players = playerDAO.getAll(roomId);
+	    if (players == null || players.isEmpty()) {
+	       gameroomDAO.deleteRoom(roomId);
+	       // 스케줄러도 종료
+	       ScheduledFuture<?> task = broadcastTasks.remove(roomId);
 	        if (task != null) task.cancel(true);
 	        latestPositions.remove(roomId);
-		}
-		
-		Map<String, Object> data = new HashMap<>();
-		data.put("userId", userId);
-		broadcastMessage("exit", roomId, data);
+	    }
 	}
 
 	@Override
